requirements.txt
# Poker Bankroll Decision System Requirements
# Core data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Scientific computing and statistics
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration and utilities
pyyaml>=6.0
